requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.7
faiss-cpu==1.7.4
numpy==1.24.3
sqlalchemy==2.0.23
python-multipart==0.0.6
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
tree-sitter==0.20.4
click==8.1.7
python-dotenv==1.0.0
tiktoken==0.5.1
pathspec==0.11.2
gitpython==3.1.40
structlog==23.2.0
